:shirt: Organize components in folders
