+++ HyruleCompendium/src/app/compendium/categories/categories.component.html
