+++ HyruleCompendium/src/app/compendium/category-card/category-card.component.html
