+++ HyruleCompendium/src/app/compendium/category-search/category-search.component.html
