+++ HyruleCompendium/src/app/compendium/creatures-details/creatures-details.component.html
