+++ HyruleCompendium/src/app/compendium/entries-list-panel/entries-list-panel.component.css
