+++ HyruleCompendium/src/app/compendium/entries-shell/entries-shell.component.html
