+++ HyruleCompendium/src/app/compendium/equipment-details/equipment-details.component.html
