+++ HyruleCompendium/src/app/compendium/material-details/material-details.component.html
