+++ HyruleCompendium/src/app/compendium/monster-details/monster-details.component.html
