+++ HyruleCompendium/src/app/compendium/treasure-details/treasure-details.component.html
